# WIP: Use pinned memory in tobytes