@kkraus14 2020-01-17
@jakirkham Have you found this faster in practice? I’m concerned about an extra host —> host memory copy.

From my perspective we should just invest the time in building a Cython `HostBuffer` class for now that wraps a `unique_ptr` around the raw pointer from `cudaMallocHost` and then implement buffer protocol and possibly `__array_interface__` on that class. Then we could either return the instance of that class directly or return a `memoryview` of the instance of that class. 

It eliminates the need for the extra host —> host memory copy and will set us up to easily replace the `unique_ptr` wrapping a raw pointer with a `unique_ptr` wrapping an eventual `host_buffer` C++ RAII class when it’s implemented in the future.

Thoughts?