@jakirkham 2020-01-17
It still needs some benchmarking. Will work on that in a moment.

I agree that is the right long term plan. Though maybe we also want to align with efforts to provide a pinned memory resource ( https://github.com/rapidsai/rmm/pull/141 ). Should provide us lighter weight wrapping too 😉

Just wanted to see how much mileage we could get out of this relatively simple (though admittedly less efficient approach).