@jakirkham 2020-01-17
Yeah the performance here is pretty abysmal. That said, it is much worse than what I would expect from a memory copy. Maybe I'm missing something?

Before:

```python
In [1]: import rmm                                                              

In [2]: rmm.reinitialize(pool_allocator=True, 
   ...:                  initial_pool_size=int(2**30))                          
Out[2]: 0

In [3]: b = rmm.DeviceBuffer(size=10)                                           

In [4]: b.tobytes()                                                             
Out[4]: b'\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00'

In [5]: %timeit b.tobytes()                                                     
10.2 µs ± 33.6 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

After:

```python
In [1]: import rmm                                                              

In [2]: rmm.reinitialize(pool_allocator=True, 
   ...:                  initial_pool_size=int(2**30))                          
Out[2]: 0

In [3]: b = rmm.DeviceBuffer(size=10)                                           

In [4]: b.tobytes()                                                             
Out[4]: b'\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00'

In [5]: %timeit b.tobytes()                                                     
1.12 ms ± 15.5 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```