@kkraus14 2020-01-18
Allocating host memory is expensive and you're doing it twice instead of once now. This is expected from my perspective.