@jakirkham 2020-01-18
> that's ms not us, that's crazy...

No kidding.

> Can you generate a quick Python profile to see if anything sticks out?

Sure I'll take a look. Thanks!